requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==6.0.4
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-filter==25.2
django-timezone-field==7.2.1
djangorestframework==3.17.1
gunicorn==25.3.0
iniconfig==2.3.0
kombu==5.6.2
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyarrow==23.0.1
Pygments==2.20.0
pytest==9.0.3
pytest-django==4.12.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.1.0
redis==7.4.0
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
